This is a first time modification to check the code is working.

+++ lib.cpp
@@ -1,21 +1,21 @@
 #include "lib.h"
 #include <stdio.h>
 #include <stdlib.h>
-#include <stdint.h>
+
 
 int* resize(int* old_vector, int old_size, int new_size)
 {
     int* new_vector = new int[new_size];
     if(new_size > old_size) // El nuevo vector es más grande. 
     {
-        for(uint8_t k = 0 ; k < old_size ; k++)
+        for(int k = 0 ; k < old_size ; k++)
         {
             new_vector[k] = old_vector[k];
         }
     }
     if(new_size < old_size) // El nuevo vector es más chico. 
     {
-        for(uint8_t k = 0 ; k < new_size ; k++)
+        for(int k = 0 ; k < new_size ; k++)
         {
             new_vector[k] = old_vector[k];
         }
@@ -28,11 +28,11 @@ int* resize(int* old_vector, int old_size, int new_size)
 int* removeItem(int* vec,int vec_size,int item_pos)
 {
     int* resulting_vector = new int[vec_size - 1];
-    for(uint8_t k = 0 ; k < item_pos ; k++)
+    for(int k = 0 ; k < item_pos ; k++)
     {
         resulting_vector[k]=vec[k];
     }
-    for(uint8_t k = item_pos + 1 ; k < vec_size - 1 ; k++)
+    for(int k = item_pos + 1 ; k < vec_size - 1 ; k++)
     {
         resulting_vector[k]=vec[k];
     }
@@ -44,12 +44,12 @@ int* removeItem(int* vec,int vec_size,int item_pos)
 int* insertItem(int* vec,int vec_size,int insert_pos,int insert_value)
 {
     int* resulting_vector = new int[vec_size + 1];
-    for(uint8_t k = 0 ; k < insert_pos ; k++)
+    for(int k = 0 ; k < insert_pos ; k++)
     {
         resulting_vector[k]=vec[k];
     }
     resulting_vector[insert_pos] = insert_value;
-    for(uint8_t k = insert_pos + 1 ; k < vec_size + 1 ; k++)
+    for(int k = insert_pos + 1 ; k < vec_size + 1 ; k++)
     {
         resulting_vector[k]=vec[k];
     }
@@ -62,13 +62,13 @@ int* concat(int* vec_left,int left_size,int* vec_right,int right_size)
 {
     int* resulting_vector = new int[left_size + right_size];
     // Copying the values of the vec_left in the resulting_vector
-    uint8_t k;
+    int k;
     for(k = 0 ; k < left_size ; k++)
     {
         resulting_vector[k]=vec_left[k];
     }
     // Copying the values of the vec_right in the resulting_vector
-    uint8_t m;
+    int m;
     for(m = 0 ; m < right_size; m++)
     {
         resulting_vector[m+k]=vec_left[m];
@@ -79,3 +79,4 @@ int* concat(int* vec_left,int left_size,int* vec_right,int right_size)
     vec_right = NULL; // Me aseguro que el puntero no quede apuntando a memoria que ya no pertenece a mi programa
     return resulting_vector;
 }
+
